+++ maplibre_gl/android/src/main/java/org/maplibre/maplibregl/MapLibreMapController.java
@@ -772,6 +772,7 @@ final class MapLibreMapController
   private boolean addHeatmapLayer(
       String layerName,
       String sourceName,
+      String sourceLayer,
       Float minZoom,
       Float maxZoom,
       String belowLayerId,
@@ -783,6 +784,9 @@ final class MapLibreMapController
     }
     HeatmapLayer layer = new HeatmapLayer(layerName, sourceName);
     layer.setProperties(properties);
+    if (sourceLayer != null) {
+      layer.setSourceLayer(sourceLayer);
+    }
     if (minZoom != null) {
       layer.setMinZoom(minZoom);
     }
@@ -1576,6 +1580,7 @@ final class MapLibreMapController
           final String sourceId = call.argument("sourceId");
           final String layerId = call.argument("layerId");
           final String belowLayerId = call.argument("belowLayerId");
+          final String sourceLayer = call.argument("sourceLayer");
           final Double minzoom = call.argument("minzoom");
           final Double maxzoom = call.argument("maxzoom");
           final PropertyValue[] properties =
@@ -1583,6 +1588,7 @@ final class MapLibreMapController
           if (!addHeatmapLayer(
               layerId,
               sourceId,
+              sourceLayer,
               minzoom != null ? minzoom.floatValue() : null,
               maxzoom != null ? maxzoom.floatValue() : null,
               belowLayerId,