+++ maplibre_gl_platform_interface/lib/src/method_channel_maplibre_gl.dart
@@ -989,6 +989,7 @@ class MapLibreMethodChannel extends MapLibrePlatform {
       'sourceId': sourceId,
       'layerId': layerId,
       'belowLayerId': belowLayerId,
+      'sourceLayer': sourceLayer,
       'minzoom': minzoom,
       'maxzoom': maxzoom,
       'properties': properties,